add fit to view
